fix: use cast in literals

+++ pyrite/math/literals/constants_literals.hpp
@@ -16,12 +16,12 @@ namespace pyrite::math::literals
  **********************/
 constexpr double operator"" _pi(long double value)
 {
-  return value * ::pyrite::math::pi<double>;
+  return static_cast<double>(value * ::pyrite::math::pi<long double>);
 }
 
 constexpr float operator"" _pif(long double value)
 {
-  return value * ::pyrite::math::pi<float>;
+  return static_cast<float>(value * ::pyrite::math::pi<long double>);
 }
 
 constexpr long double operator"" _pil(long double value)
@@ -31,18 +31,18 @@ constexpr long double operator"" _pil(long double value)
 
 constexpr double operator"" _pi(unsigned long long value)
 {
-  return value * ::pyrite::math::pi<double>;
+  return static_cast<double>(value * ::pyrite::math::pi<long double>);
 }
 
 constexpr float operator"" _pif(unsigned long long value)
 {
-  return value * ::pyrite::math::pi<float>;
+  return static_cast<float>(value * ::pyrite::math::pi<long double>);
 }
 
 constexpr long double operator"" _pil(unsigned long long value)
 {
   return value * ::pyrite::math::pi<long double>;
 }
-} // namespace pyrite::literals::math_literals
+} // namespace pyrite::math::literals
 
 #endif // PYRITE_MATH_CONSTANTS_LITERALS_HPP
